Move checksums to util dir
